requirements.txt
beautifulsoup4
numpy
python-Levenshtein
requests
selenium
yt-dlp
